refactor: relocate testcase definitions
