+++ httprunner/__init__.py
@@ -1,8 +1,9 @@
 __version__ = "3.0.7"
 __description__ = "One-stop solution for HTTP(S) testing."
 
-from httprunner.runner import HttpRunner, Config, Step, Request
+from httprunner.runner import HttpRunner
 from httprunner.schema import TConfig, TStep
+from httprunner.testcase import Config, Step, Request
 
 __all__ = [
     "__version__",
